Move FeatureSet to its own file
